wip: more infisical changes, misc cleanup, remove failure check in favor of just checking everything
