+++ .github/workflows/bot_image.yml
@@ -69,6 +69,15 @@ jobs:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
       -
+        name: Set NODE_ENV
+        id: node_env
+        run: |
+          if [[ "${{ github.ref_name }}" == "master" ]]; then
+            echo "value=production" >> $GITHUB_OUTPUT
+          else
+            echo "value=development" >> $GITHUB_OUTPUT
+          fi
+      -
         name: Build and push
         uses: docker/build-push-action@v5
         with:
@@ -76,3 +85,5 @@ jobs:
           file: services/bot/Dockerfile
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
+          build-args: |
+            NODE_ENV=${{ steps.node_env.outputs.value }}
