+++ .infisical.json
@@ -1,0 +1,5 @@
+{
+    "workspaceId": "1e4a9be2-aaf8-4e0b-be53-185dac145515",
+    "defaultEnvironment": "",
+    "gitBranchToEnvironmentMapping": null
+}
