+++ packages/store/drizzle/0009_happy_omega_red.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "metadata" ADD COLUMN "start" numeric NOT NULL;--> statement-breakpoint
+ALTER TABLE "metadata" DROP COLUMN "oldestUnansweredQuestion";
