+++ packages/store/drizzle/0010_keen_komodo.sql
@@ -1,0 +1,1 @@
+DROP TABLE "failures" CASCADE;
