+++ packages/store/drizzle/0011_new_scarlet_witch.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "metadata" ALTER COLUMN "start" SET DEFAULT 1;
