+++ packages/store/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./database";
-export * from "./resources";
 export * from "./storage";
 export * from "./schema_types";
