--- packages/store/src/resources.ts
@@ -1,1 +1,0 @@
-export const QNAPLUS_BUCKET = "qnaplus";
