+++ packages/typescript-config/base.json
@@ -10,6 +10,7 @@
 		"esModuleInterop": true,
 		"skipLibCheck": true,
 		"moduleResolution": "node16",
-		"composite": true
+		"composite": true,
+        "types": ["node"]
 	}
 }
