+++ services/updater/Dockerfile
@@ -19,13 +19,14 @@ COPY . .
 
 FROM base AS build
 
+ARG NODE_ENV=production
 # Hoisted mode so pnpm physically copies platform-specific optional binaries
 # instead of leaving them as symlinks that break across the cache-mount boundary.
 RUN printf '\nnodeLinker: hoisted\n' >> pnpm-workspace.yaml
 RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm i --frozen-lockfile
 RUN pnpm build
 
-ENV NODE_ENV=production
+ENV NODE_ENV=$NODE_ENV
 RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm --filter @qnaplus/updater --prod deploy out \
 && mv /app/.npmrc out \
 && mv /app/services/start.sh out
@@ -37,8 +38,11 @@ RUN npm install -g @infisical/cli@0.43.84
 
 # =======================================================
 
-FROM node:22.14.0-bookworm-slim AS prod
+FROM node:22.14.0-bookworm-slim
 RUN npm install -g @infisical/cli@0.43.84
+
+ARG NODE_ENV=production
+ENV NODE_ENV=$NODE_ENV
 
 COPY --from=infisical-cli /usr/local/bin/infisical /usr/local/bin/infisical
 COPY --from=build /app/out /app
